Add module Roles

+++ app/Modules/Admin/Resources/Views/common/controls/destroy.blade.php
@@ -1,5 +1,7 @@
-{!! Form::open(['route' => [$routePrefix.'destroy', 'id'=>$id], 'method' => 'delete']) !!}
-<button type="submit" class="btn btn-danger btn-sm" title="Удалить">
-    <i class="glyphicon glyphicon-trash"></i>
-</button>
-{!! Form::close() !!}
+@if(Auth::guard('admin')->user()->canDelete())
+    {!! Form::open(['route' => [$routePrefix.'destroy', 'id'=>$id], 'method' => 'delete']) !!}
+    <button type="submit" class="btn btn-danger btn-sm" title="Удалить">
+        <i class="glyphicon glyphicon-trash"></i>
+    </button>
+    {!! Form::close() !!}
+@endif
